--- src/components/InputBlock/index.js
@@ -1,3 +1,0 @@
-import InputBlock from './InputBlock';
-
-export default InputBlock;
