+++ src/components/InputBlockContainer/index.js
@@ -1,0 +1,3 @@
+import InputBlockContainer from './InputBlockContainer';
+
+export default InputBlockContainer;
