+++ src/components/InputBlocks/index.js
@@ -1,0 +1,4 @@
+// UPLOAD INPUT BLOCK
+import UploadInputBlock from './UploadInputBlock';
+
+export { UploadInputBlock };
